Fix: rename AdvancedResourceGridItem folder
